+++ citygml4j-core/src/main/java/org/citygml4j/core/model/core/ImplicitGeometry.java
@@ -17,6 +17,7 @@ import org.xmlobjects.gml.util.Matrices;
 import org.xmlobjects.gml.util.matrix.Matrix;
 import org.xmlobjects.model.ChildList;
 
+import java.util.Arrays;
 import java.util.List;
 
 public class ImplicitGeometry extends AbstractGML implements CityGMLObject, VisitableObject {
@@ -118,8 +119,19 @@ public class ImplicitGeometry extends AbstractGML implements CityGMLObject, Visi
                     matrix.set(2, 3, matrix.get(2, 3) + point.get(2));
                 }
 
-                envelope.include(Matrices.transform3D(relative.getLowerCorner(), matrix));
-                envelope.include(Matrices.transform3D(relative.getUpperCorner(), matrix));
+                List<Double> lowerCorner = relative.getLowerCorner().toCoordinateList3D();
+                List<Double> upperCorner = relative.getUpperCorner().toCoordinateList3D();
+                for (int x = 0; x < 2; x++) {
+                    for (int y = 0; y < 2; y++) {
+                        for (int z = 0; z < 2; z++) {
+                            List<Double> corner = Arrays.asList(
+                                    x == 0 ? lowerCorner.get(0) : upperCorner.get(0),
+                                    y == 0 ? lowerCorner.get(1) : upperCorner.get(1),
+                                    z == 0 ? lowerCorner.get(2) : upperCorner.get(2));
+                            envelope.include(Matrices.transform3D(corner, matrix));
+                        }
+                    }
+                }
             }
         } else if (referencePoint != null
                 && referencePoint.getObject() != null) {